DATAMONGO-1819 - Do not create PersistentEntities for simple types contributed by a converter.
We now check for registered simple types in MongoMappingContext.shouldCreatePersistentEntityFor(…) to prevent PersistentEntity creation for unwanted types. Previously, we only checked against MongoSimpleTypes.HOLDER instead of the simple types contributed by converters.

This change impacts aggregate root type use with top-level converters: These types are no longer represented with a PersistentEntity which removes mapping details and these types cannot be used anymore with a repository and require a collection name if used with the Template API.